Defect fixes: Defect 1: In the Signup screen please provide specific description for password needs to be whether numeric,alphanumeric or adding special characters.
Defect 2:The error message after submitting is at a position which is not visible to user please change its position,.

Defect 3. When Signing up it showed an error when SUBMITTING signup it says "options is something!" Error.
Please fix that -- Although after entering 5 times different password it didn't even show me a account succesfully created or after creation didn't took me to the main screen it was showing the option error
the password i entered was:-
 i. Pavansin
 ii. Pavansin.io
 iii. Ankush@1234@
 iv.ABCD@1234abcd (this one got saved and created an account for me)

Defect 4: So i decided to reload the page and try to make the account again as i was confused whether the account has been created or not so i tried and it was give me duplicate key error with some syntax in the error message so handle that.

Defect 5: Now went  to login screen and logged in after logging in the 1st i pressed was interested then a screen showed me for a second that "No More Profiles Available - You have interacted with all the profiles. Please check back later for new connectio" but after a second a new card comes into play so please fix that also.

Defect 6: Now went to the profile section where i can see a form which i do not want to see i want to see my info also. Validation for the form is missing and can paste anything in the url even if it doesnt contain any .com,.in something,the about has not character limit same as others also the limit need to be specified

Defect 7: Payment failure also leads to premium subscription.